requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.1.1
django-ninja==1.3.0
et-xmlfile==1.1.0
gunicorn==23.0.0
h11==0.14.0
openpyxl==3.1.5
orjson==3.10.7
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.1
whitenoise==6.7.0
